phase_rare: haplotype_writer: remove missing flag on imputed variants
Sparse genotypes for imputed variants have the missing flag until the
end, in the writeHaplotypeVCF() function only the alleles and probability
are carried over to the VCF record, not the "missing flag".

However with writeHaplotyeXCF() the sparse genotypes are directly written
in the binary file and therefore carry over any flag. Therefore remove
the missing flag in the sparse genotypes that still have it (i.e.,
imputed genotypes).

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@unil.ch>